EE361 2019 midterm 1 Q1 Modelling
2D modelling: circuit magnetostatic, PM magnetostatic and transient.
Demagnetization analyzed for middle point of the PM, for several variations.

MATLAB code: BH curves given for VAC633AP20degC is in terms of B(T) vs. H(kOe). Code changes the H(kOe) to H(A/m) and provides a .tab file suitable for importing to ANSYS as a rectangular plot.

.py files: exemplary script files extracted from ANSYS.